aws/application/Api/requirements.txt
python-json-logger==0.1.11
pyyaml==5.1.2
context-log==0.1.2
boto3==1.10.14
flask==1.1.1
aws-wsgi==0.2.3
